+++ src/main/java/com/piseth/java/school/schoolManagement/spec/MonthlyScoreFilter.java
@@ -9,8 +9,8 @@ public class MonthlyScoreFilter {
 	private List<Long> studentIds;
 	private List<Long> subjectIds;
 	private Long  studentId;
-	private Short year;
-	private Short month;
-	private Short grade;
+	private Integer year;
+	private Integer month;
+	private Integer grade;
 	private String className;
 }
